Organize ImageView folder
- Add `components` folder for components
- Add `hooks` folder for hooks
- Move "libraries" to `lib` folder
