+++ packages/studio-base/src/panels/ImageView/components/TopicTimestamp.tsx
