+++ packages/studio-base/src/panels/ImageView/lib/HitmapRenderContext.ts
