+++ packages/studio-base/src/panels/ImageView/lib/PinholeCameraModel.ts
