+++ packages/studio-base/src/panels/ImageView/lib/decodings.test.ts
