+++ packages/studio-base/src/panels/ImageView/lib/decodings.ts
@@ -11,7 +11,7 @@
 //   found at http://www.apache.org/licenses/LICENSE-2.0
 //   You may not use this file except in compliance with the License.
 
-import { RenderOptions } from "./types";
+import { RenderOptions } from "../types";
 
 export function decodeYUV(
   yuv: Int8Array,
