+++ packages/studio-base/src/panels/ImageView/lib/util.test.ts
